Added Solution for Exercises 4.15 - 4.23 of Part4d

+++ Part4/bloglist/controller/Blogs.js
@@ -1,9 +1,12 @@
+require('dotenv').config()
+const middleware = require('../utils/middleware')
 const BlogsRouter = require('express').Router()
-const { request } = require('../app')
-const Blog = require('../models/blog')
+const Blog = require('../models/blogs')
+const User = require('../models/users')
+const jwt = require('jsonWebToken')
 
 BlogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.status(200).json(blogs)
 })
 
@@ -18,21 +21,45 @@ BlogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-BlogsRouter.post('/', async (request, response) => {
+BlogsRouter.post('/', middleware.userExtractor ,async (request, response) => {
+
   if(!request.body.title || !request.body.url ){
     return response.status(400).json({error: "title and url required"})
   }
-  else{
-    const blog = new Blog(request.body)
-    await blog.save()
-    return response.status(201).json(blog)
-  }
+  const body = request.body
+  const user = request.user
   
-
+  const blog = new Blog({
+    url: body.url,
+    title: body.title,
+    author: body.author,
+    user: user._id,
+  })
+  const newBlog = await blog.save()
+  user.blogs = user.blogs.concat(newBlog._id)
+  await user.save()
+  
+  return response.status(201).json(newBlog)
 })
 
-BlogsRouter.delete('/:id', async (request, response) => {
-  const deleteBlog = await Blog.findByIdAndDelete(request.params.id)
+BlogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+
+  if(!blog){
+    return response.status(404).json({
+      error: `blog id of ${request.params.id} not found`
+    })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if(decodedToken.id !== blog.user.toString()){
+    return response.status(403).json({
+      error: "You're forbidden to delete this blog"
+    })
+  }
+
+  await Blog.findByIdAndDelete(request.params.id)
   return response.status(204).end()
 })
